fix(bindings): validate and bound public audio input on the WASM and Node surfaces
Preflight caller-owned Float32Array lengths against a wasm32 element budget
before any heap copy, rejecting non-finite, fractional, or oversized lengths
and matrix shapes whose element count would overflow. Reject unknown pan
modes and out-of-range pan ordinals instead of silently falling back to
balance, and propagate one-shot strip setter failures rather than swallowing
them. Range-check the metering sample rate, clamp streaming block and tail
drain sizes to the prepared block size, and keep the borrowed sidechain
pointer array alive across updates so a failed update cannot leave a dangling
key. Extend the shared conformance corpus with mix-option rejection cases and
exercise them on every surface.

+++ src/wasm/bindings/mixing/mixing_wasm.h
@@ -195,7 +195,7 @@ class MixerWasm {
 
   // Drains delayed/tail audio by processing a zero-input block of num_samples
   // frames. Returns { left, right, sampleRate } mirroring processStereo.
-  val drainTailStereo(size_t num_samples);
+  val drainTailStereo(double num_samples);
 
  private:
   static void checkStripError(SonareError err, const char* what);
